+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -34,6 +34,7 @@ import {
 import {
   DEFAULT_CURRENCY,
   DEFAULT_DATE_RANGE,
+  TAG_ID_DRAFT,
   TAG_ID_EMERGENCY_FUND,
   TAG_ID_EXCLUDE_FROM_ANALYSIS,
   UNKNOWN_KEY
@@ -43,6 +44,7 @@ import {
   getAssetProfileIdentifier,
   getSum,
   isAccountExcluded,
+  isDraftActivity,
   parseDate
 } from '@ghostfolio/common/helper';
 import {
@@ -174,7 +176,19 @@ export class PortfolioService {
       this.accountService.accounts({
         where,
         include: {
-          activities: { include: { SymbolProfile: true } },
+          activities: {
+            include: {
+              SymbolProfile: true,
+              tags: {
+                select: {
+                  id: true
+                },
+                where: {
+                  id: TAG_ID_DRAFT
+                }
+              }
+            }
+          },
           platform: true,
           tags: true
         },
@@ -200,12 +214,18 @@ export class PortfolioService {
         for (const {
           currency,
           date,
-          isDraft,
           quantity,
           SymbolProfile,
+          tags,
           type,
           unitPrice
         } of account.activities) {
+          activitiesCount += 1;
+
+          if (isDraftActivity({ tags })) {
+            continue;
+          }
+
           switch (type) {
             case ActivityType.DIVIDEND:
               dividendInBaseCurrency +=
@@ -225,10 +245,6 @@ export class PortfolioService {
                   date
                 )) ?? 0;
               break;
-          }
-
-          if (!isDraft) {
-            activitiesCount += 1;
           }
         }
 
@@ -2101,8 +2117,8 @@ export class PortfolioService {
   }) {
     return getSum(
       activities
-        .filter(({ isDraft, type }) => {
-          return isDraft === false && type === activityType;
+        .filter((activity) => {
+          return !isDraftActivity(activity) && activity.type === activityType;
         })
         .map(({ assetProfile, currency, quantity, unitPrice }) => {
           return new Big(