+++ apps/client/src/app/pages/accounts/accounts-page.component.ts
@@ -1,17 +1,9 @@
-import { GfAccountDetailDialogComponent } from '@ghostfolio/client/components/account-detail-dialog/account-detail-dialog.component';
-import {
-  AccountDetailDialogParams,
-  AccountDetailDialogResult
-} from '@ghostfolio/client/components/account-detail-dialog/interfaces/interfaces';
 import { ImpersonationStorageService } from '@ghostfolio/client/services/impersonation-storage.service';
 import { UserService } from '@ghostfolio/client/services/user/user.service';
-import {
-  CreateAccountDto,
-  TransferBalanceDto,
-  UpdateAccountDto
-} from '@ghostfolio/common/dtos';
+import { TransferBalanceDto } from '@ghostfolio/common/dtos';
 import { User } from '@ghostfolio/common/interfaces';
 import { hasPermission, permissions } from '@ghostfolio/common/permissions';
+import { internalRoutes } from '@ghostfolio/common/routes/routes';
 import { AccountWithValue } from '@ghostfolio/common/types';
 import { GfAccountsTableComponent } from '@ghostfolio/ui/accounts-table';
 import { GfFabComponent } from '@ghostfolio/ui/fab';
@@ -30,13 +22,10 @@ import {
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
-import { Tag } from '@prisma/client';
 import { DeviceDetectorService } from 'ngx-device-detector';
 import { EMPTY } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 
-import { GfCreateOrUpdateAccountDialogComponent } from './create-or-update-account-dialog/create-or-update-account-dialog.component';
-import { CreateOrUpdateAccountDialogParams } from './create-or-update-account-dialog/interfaces/interfaces';
 import { TransferBalanceDialogParams } from './transfer-balance/interfaces/interfaces';
 import { GfTransferBalanceDialogComponent } from './transfer-balance/transfer-balance-dialog.component';
 
@@ -54,9 +43,12 @@ export class GfAccountsPageComponent implements OnInit {
   protected hasPermissionToCreateAccount: boolean;
   protected hasPermissionToUpdateAccount: boolean;
   protected impersonationId: string | null;
+  protected readonly internalRoutes = internalRoutes;
   protected totalBalanceInBaseCurrency = 0;
   protected totalValueInBaseCurrency = 0;
   protected user: User;
+
+  private isInitialFetch = true;
 
   private readonly deviceType = computed(
     () => this.deviceDetectorService.deviceInfo().deviceType
@@ -79,26 +71,7 @@ export class GfAccountsPageComponent implements OnInit {
     this.route.queryParams
       .pipe(takeUntilDestroyed(this.destroyRef))
       .subscribe((params) => {
-        if (params['accountId'] && params['accountDetailDialog']) {
-          this.openAccountDetailDialog(params['accountId']);
-        } else if (
-          params['createDialog'] &&
-          this.hasPermissionToCreateAccount
-        ) {
-          this.openCreateAccountDialog();
-        } else if (params['editDialog']) {
-          if (this.accounts) {
-            const account = this.accounts.find(({ id }) => {
-              return id === params['accountId'];
-            });
-
-            if (account) {
-              this.openUpdateAccountDialog(account);
-            }
-          } else {
-            this.router.navigate(['.'], { relativeTo: this.route });
-          }
-        } else if (params['transferBalanceDialog']) {
+        if (params['transferBalanceDialog']) {
           this.openTransferBalanceDialog();
         }
       });
@@ -130,12 +103,12 @@ export class GfAccountsPageComponent implements OnInit {
             this.user.permissions,
             permissions.updateAccount
           );
+
+          this.fetchAccounts();
         }
 
         this.changeDetectorRef.markForCheck();
       });
-
-    this.fetchAccounts();
   }
 
   protected onDeleteAccount(aId: string) {
@@ -149,20 +122,12 @@ export class GfAccountsPageComponent implements OnInit {
           .get(true)
           .pipe(takeUntilDestroyed(this.destroyRef))
           .subscribe();
-
-        this.fetchAccounts();
       });
   }
 
   protected onTransferBalance() {
     this.router.navigate([], {
       queryParams: { transferBalanceDialog: true }
-    });
-  }
-
-  protected onUpdateAccount(aAccount: AccountWithValue) {
-    this.router.navigate([], {
-      queryParams: { accountId: aAccount.id, editDialog: true }
     });
   }
 
@@ -182,149 +147,21 @@ export class GfAccountsPageComponent implements OnInit {
           this.totalBalanceInBaseCurrency = totalBalanceInBaseCurrency;
           this.totalValueInBaseCurrency = totalValueInBaseCurrency;
 
-          if (this.accounts?.length <= 0) {
-            this.router.navigate([], { queryParams: { createDialog: true } });
+          if (
+            this.accounts?.length <= 0 &&
+            this.hasPermissionToCreateAccount &&
+            this.isInitialFetch
+          ) {
+            void this.router.navigate(
+              internalRoutes.accounts.subRoutes.create.routerLink
+            );
           }
+
+          this.isInitialFetch = false;
 
           this.changeDetectorRef.markForCheck();
         }
       );
-  }
-
-  private openUpdateAccountDialog({
-    balance,
-    comment,
-    currency,
-    id,
-    name,
-    platformId,
-    tags
-  }: AccountWithValue & { tags?: Tag[] }) {
-    const dialogRef = this.dialog.open<
-      GfCreateOrUpdateAccountDialogComponent,
-      CreateOrUpdateAccountDialogParams
-    >(GfCreateOrUpdateAccountDialogComponent, {
-      data: {
-        account: {
-          balance,
-          comment,
-          currency,
-          id,
-          name,
-          platformId,
-          tags
-        },
-        user: this.user
-      },
-      height: this.deviceType() === 'mobile' ? '98vh' : '80vh',
-      width: this.deviceType() === 'mobile' ? '100vw' : '50rem'
-    });
-
-    dialogRef
-      .afterClosed()
-      .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe((account: UpdateAccountDto | null) => {
-        if (account) {
-          this.reset();
-
-          this.dataService
-            .putAccount(account)
-            .pipe(takeUntilDestroyed(this.destroyRef))
-            .subscribe(() => {
-              this.userService
-                .get(true)
-                .pipe(takeUntilDestroyed(this.destroyRef))
-                .subscribe();
-
-              this.fetchAccounts();
-            });
-
-          this.changeDetectorRef.markForCheck();
-        }
-
-        this.router.navigate(['.'], { relativeTo: this.route });
-      });
-  }
-
-  private openAccountDetailDialog(aAccountId: string) {
-    const dialogRef = this.dialog.open<
-      GfAccountDetailDialogComponent,
-      AccountDetailDialogParams,
-      AccountDetailDialogResult
-    >(GfAccountDetailDialogComponent, {
-      autoFocus: false,
-      data: {
-        accountId: aAccountId,
-        deviceType: this.deviceType(),
-        hasPermissionToCreateActivity:
-          !this.hasImpersonationId &&
-          hasPermission(this.user?.permissions, permissions.createActivity) &&
-          !this.user?.settings?.isRestrictedView,
-        impersonationId: this.impersonationId
-      },
-      height: this.deviceType() === 'mobile' ? '98vh' : '80vh',
-      width: this.deviceType() === 'mobile' ? '100vw' : '50rem'
-    });
-
-    dialogRef
-      .afterClosed()
-      .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe((result) => {
-        if (result?.isNavigating) {
-          return;
-        }
-
-        this.fetchAccounts();
-
-        this.router.navigate(['.'], { relativeTo: this.route });
-      });
-  }
-
-  private openCreateAccountDialog() {
-    const dialogRef = this.dialog.open<
-      GfCreateOrUpdateAccountDialogComponent,
-      CreateOrUpdateAccountDialogParams
-    >(GfCreateOrUpdateAccountDialogComponent, {
-      data: {
-        account: {
-          balance: 0,
-          comment: null,
-          currency: this.user?.settings?.baseCurrency ?? null,
-          id: null,
-          name: null,
-          platformId: null,
-          tags: []
-        },
-        user: this.user
-      } satisfies CreateOrUpdateAccountDialogParams,
-      height: this.deviceType() === 'mobile' ? '98vh' : '80vh',
-      width: this.deviceType() === 'mobile' ? '100vw' : '50rem'
-    });
-
-    dialogRef
-      .afterClosed()
-      .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe((account: CreateAccountDto | null) => {
-        if (account) {
-          this.reset();
-
-          this.dataService
-            .postAccount(account)
-            .pipe(takeUntilDestroyed(this.destroyRef))
-            .subscribe(() => {
-              this.userService
-                .get(true)
-                .pipe(takeUntilDestroyed(this.destroyRef))
-                .subscribe();
-
-              this.fetchAccounts();
-            });
-
-          this.changeDetectorRef.markForCheck();
-        }
-
-        this.router.navigate(['.'], { relativeTo: this.route });
-      });
   }
 
   private openTransferBalanceDialog() {