+++ libs/ui/src/lib/i18n.ts
@@ -18,6 +18,7 @@ const locales = {
   DATA_IMPORT_AND_EXPORT_TOOLTIP_BASIC: $localize`Switch to Ghostfolio Premium or Ghostfolio Open Source easily`,
   DATA_IMPORT_AND_EXPORT_TOOLTIP_OSS: $localize`Switch to Ghostfolio Premium easily`,
   DATA_SOURCE: $localize`Data Source`,
+  DRAFT: $localize`Draft`,
   EMERGENCY_FUND: $localize`Emergency Fund`,
   EXCLUDE_FROM_ANALYSIS: $localize`Exclude from Analysis`,
   Global: $localize`Global`,